Update NuGet packages to fix Dependabot security alerts
- Magick.NET 13.5.0 → 14.10.2 (fixes 18 ImageMagick CVEs)
- Azure.Storage.Blobs 12.19.1 → 12.27.0
- Microsoft.Extensions.Configuration 8.0.x → 9.0.x (matches net9.0 target)
- Fix breaking API changes (uint types in Magick.NET, GetBlobsOptions in Azure SDK)

+++ tools/src/ImageProcessor/Models/LocalImage.cs
@@ -25,9 +25,9 @@ public class LocalImage(string imagePath, string relativeImagePath) : IDisposabl
         }
     }
 
-    public int WidthPixels => ImageData.Width;
+    public int WidthPixels => (int)ImageData.Width;
 
-    public int HeightPixels => ImageData.Height;
+    public int HeightPixels => (int)ImageData.Height;
 
     public MagickFormat ImageFormat => ImageData.Format;
 
@@ -43,7 +43,7 @@ public class LocalImage(string imagePath, string relativeImagePath) : IDisposabl
     public string ResizeAndSaveToTemp(WebImageFormat format, int width, int? quality = null)
     {
         var newImage = ImageData.Clone();
-        newImage.Resize(width, 0);
+        newImage.Resize((uint)width, 0);
         string extension;
 
         // Determine the format and set the extension
@@ -72,7 +72,7 @@ public class LocalImage(string imagePath, string relativeImagePath) : IDisposabl
         {
             if (quality is not null && quality.Value > 0 && quality.Value <= 100)
             {
-                newImage.Quality = quality.Value;
+                newImage.Quality = (uint)quality.Value;
             }
         }
